# Conflicts:
#	Controllers/AuthController.cs
#	Controllers/AuthSellerController.cs
#	Controllers/SellersController.cs
#	Service/SmsSend.cs

+++ Models/Merchandise.cs
@@ -23,7 +23,7 @@ namespace EBook.Models
 
         public string Description { set; get; }
 
-        public int Price { set; get; }
+        public double Price { set; get; }
